requirements.txt
# Streamlit Frontend Requirements (API client only)
# Backend (Phase 4) runs separately. Install: pip install -r requirements.txt
# Run: streamlit run streamlit_app.py

streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0
